Adding changes to orm for addition of sql statement directly

+++ taskcontrol/orm.py
@@ -1,6 +1,3 @@
-# decide package
-import sqlite3
-
 # Inherit shared and logging
 
 # RESOURCES for later
@@ -20,15 +17,20 @@ class SQLORM(SQLBase):
 
     def create(self, conn, options):
         try:
-            sql = """INSERT INTO """ + str(options.get("table"))
-            for i in options.get("columns"):
-                sql += """ (""" + str(i) + """, """
-            sql += """) VALUES ( """
+            sql = options.get("sql")
+            if type(sql) == str and len(sql) > 0:
+                pass
+            else:
+                sql = """INSERT INTO """ + str(options.get("table"))
+                for i in options.get("columns"):
+                    sql += """ (""" + str(i) + """, """
+                sql += """) VALUES ( """
 
-            for j in options.get("values"):
-                sql += str(j) + """, """
+                for j in options.get("values"):
+                    sql += str(j) + """, """
 
-            sql += """);"""
+                sql += """);"""
+
             conn.execute(sql)
             conn.commit()
             print(options.get("table"), " created successfully")
@@ -38,33 +40,37 @@ class SQLORM(SQLBase):
 
     def find(self, conn, options):
         try:
-            sql = """SELECT """
-            for i in options.get("columns"):
-                sql += str(i) + """, """
-            sql += """ FROM """ + str(options.get("table")) + """ WHERE """
-            sql += options.get("conditions")
+            sql = options.get("sql")
+            if type(sql) == str and len(sql) > 0:
+                pass
+            else:
+                sql = """SELECT """
+                for i in options.get("columns"):
+                    sql += str(i) + """, """
+                sql += """ FROM """ + str(options.get("table")) + """ WHERE """
+                sql += options.get("conditions")
 
-            # TODO
-            # Extend later with all filters, joins, and nested
-            # Priority get this working
+                # TODO
+                # Extend later with all filters, joins, and nested
+                # Priority get this working
 
-            # filters = options.get("filters")
-            # if type(filters) == str:
-            #     sql += filters + """;"""
-            # elif type(filters) == dict:
-            #     # TODO:
-            #     # Make this nested for joins, nested statements, and with all operators
-            #     # Currently keeping it only for string and single statements
-            #     #
-            #     # Not priority
-            #     # Reason:
-            #     # Let users work on their own DB based systems
-            #     #       for other activities in plugin by extending
-            #     # Handle only user authentication for small
-            #     #       apps and let users scale with their db
-            #     # Put SQLITE and Pickle data into memory for every instance
-            #     # Make writes to memory and DB to persist
-            #     pass
+                # filters = options.get("filters")
+                # if type(filters) == str:
+                #     sql += filters + """;"""
+                # elif type(filters) == dict:
+                #     # TODO:
+                #     # Make this nested for joins, nested statements, and with all operators
+                #     # Currently keeping it only for string and single statements
+                #     #
+                #     # Not priority
+                #     # Reason:
+                #     # Let users work on their own DB based systems
+                #     #       for other activities in plugin by extending
+                #     # Handle only user authentication for small
+                #     #       apps and let users scale with their db
+                #     # Put SQLITE and Pickle data into memory for every instance
+                #     # Make writes to memory and DB to persist
+                #     pass
 
             conn.execute(sql)
             conn.commit()
@@ -75,20 +81,24 @@ class SQLORM(SQLBase):
 
     def update(self, conn, options):
         try:
-            sql = """UPDATE """ + options.get("table")
-            sql += """ SET """
-            # UPDATE STATEMENTS
-            sql += options.get("statements")
-            sql += """ WHERE """
-            # UPDATE CONDITION STATEMENTS
+            sql = options.get("sql")
+            if type(sql) == str and len(sql) > 0:
+                pass
+            else:
+                sql = """UPDATE """ + options.get("table")
+                sql += """ SET """
+                # UPDATE STATEMENTS
+                sql += options.get("statements")
+                sql += """ WHERE """
+                # UPDATE CONDITION STATEMENTS
 
-            # Applying conditions
-            con = options.get("conditions")
-            if con and type(con) == str:
-                sql += con
+                # Applying conditions
+                con = options.get("conditions")
+                if con and type(con) == str:
+                    sql += con
 
-            # filters = options.get("filters")
-            sql += """;"""
+                # filters = options.get("filters")
+                sql += """;"""
             conn.execute(sql)
             conn.commit()
             print(options.get("table"), " updated successfully")
@@ -98,34 +108,37 @@ class SQLORM(SQLBase):
 
     def delete(self, conn, options):
         try:
-            sql = """DELETE FROM """
-            sql += str(options.get("table")) + """ WHERE """
+            sql = options.get("sql")
+            if type(sql) == str and len(sql) > 0:
+                pass
+            else:
+                sql = """DELETE FROM """
+                sql += str(options.get("table")) + """ WHERE """
 
-            # Applying conditions
-            con = options.get("conditions")
-            if con and type(con) == str:
-                sql += con
-            # filters = options.get("filters")
-            # if type(filters) == str:
-            #     sql += filters + """;"""
-            # elif type(filters) == dict:
-            #     # TODO:
-            #     # Make this nested for joins, nested statements, and with all operators
-            #     # Currently keeping it only for string and single statements
-            #     #
-            #     # Not priority
-            #     # Reason:
-            #     # Let users work on their own DB based systems
-            #     #       for other activities in plugin by extending
-            #     # Handle only user authentication for small
-            #     #       apps and let users scale with their db
-            #     # Put SQLITE and Pickle data into memory for every instance
-            #     # Make writes to memory and DB to persist
-            #     pass
+                # Applying conditions
+                con = options.get("conditions")
+                if con and type(con) == str:
+                    sql += con
+                # filters = options.get("filters")
+                # if type(filters) == str:
+                #     sql += filters + """;"""
+                # elif type(filters) == dict:
+                #     # TODO:
+                #     # Make this nested for joins, nested statements, and with all operators
+                #     # Currently keeping it only for string and single statements
+                #     #
+                #     # Not priority
+                #     # Reason:
+                #     # Let users work on their own DB based systems
+                #     #       for other activities in plugin by extending
+                #     # Handle only user authentication for small
+                #     #       apps and let users scale with their db
+                #     # Put SQLITE and Pickle data into memory for every instance
+                #     # Make writes to memory and DB to persist
+                #     pass
             conn.execute(sql)
             conn.commit()
             print(options.get("table"), " deleted successfully")
         except Exception as e:
             raise Exception("Error with options provided", e)
         return True
-
